+++ packages/react-native-audio-api/ios/audioapi/ios/core/utils/AudioDecoder.mm
@@ -41,8 +41,10 @@ std::vector<float> AudioDecoder::readAllPcmFrames(ma_decoder &decoder, int outpu
   return buffer;
 }
 
-std::shared_ptr<AudioBuffer>
-AudioDecoder::makeAudioBufferFromFloatBuffer(const std::vector<float> &buffer, int outputSampleRate, int outputChannels)
+std::shared_ptr<AudioBuffer> AudioDecoder::makeAudioBufferFromFloatBuffer(
+    const std::vector<float> &buffer,
+    float outputSampleRate,
+    int outputChannels)
 {
   if (buffer.empty()) {
     return nullptr;
@@ -85,7 +87,7 @@ std::shared_ptr<AudioBuffer> AudioDecoder::decodeWithFilePath(const std::string 
     return nullptr;
   }
 
-  auto outputSampleRate = static_cast<int>(decoder.outputSampleRate);
+  auto outputSampleRate = static_cast<float>(decoder.outputSampleRate);
   auto outputChannels = static_cast<int>(decoder.outputChannels);
 
   buffer = readAllPcmFrames(decoder, outputChannels);
@@ -120,7 +122,7 @@ std::shared_ptr<AudioBuffer> AudioDecoder::decodeWithMemoryBlock(const void *dat
     return nullptr;
   }
 
-  auto outputSampleRate = static_cast<int>(decoder.outputSampleRate);
+  auto outputSampleRate = static_cast<float>(decoder.outputSampleRate);
   auto outputChannels = static_cast<int>(decoder.outputChannels);
 
   buffer = readAllPcmFrames(decoder, outputChannels);
